+++ cypress/integration/WidgetsForEnrollmentPages/WidgetsForEnrollmentDashboard/index.js
@@ -1,6 +1,7 @@
 import '../sharedSteps';
 import '../WidgetEnrollment';
 import '../WidgetProfile';
+import '../WidgetEnrollmentComment';
 
 beforeEach(() => {
     cy.loginThroughForm();