+++ src/core_modules/capture-core/components/Pages/EnrollmentAddEvent/EnrollmentAddEventPage.types.js
@@ -18,5 +18,6 @@ export type Props = {|
     rulesExecutionDependencies: Object,
     pageFailure: boolean,
     ready: boolean,
+    widgetReducerName: string,
     ...CssClasses,
 |};