+++ src/core_modules/capture-core/components/Pages/common/EnrollmentOverviewDomain/enrollment.actions.js
@@ -5,8 +5,12 @@ import { actionCreator } from '../../../../actions/actions.utils';
 export const enrollmentSiteActionTypes = {
     COMMON_ENROLLMENT_SITE_DATA_SET: 'EnrollmentSite.SetCommonData',
     UPDATE_ENROLLMENT_EVENTS: 'Enrollment.UpdateEnrollmentEvents',
+    UPDATE_ENROLLMENT_EVENTS_WITHOUT_ID: 'Enrollment.UpdateEnrollmentEventsWithoutId',
     ROLLBACK_ENROLLMENT_EVENT: 'Enrollment.RollbackEnrollmentEvent',
+    ROLLBACK_ENROLLMENT_EVENT_WITHOUT_ID: 'Enrollment.RollbackEnrollmentEventWithoutId',
     COMMIT_ENROLLMENT_EVENT: 'Enrollment.CommitEnrollmentEvent',
+    COMMIT_ENROLLMENT_EVENT_WITHOUT_ID: 'Enrollment.CommitEnrollmentEventWithoutId',
+    SAVE_FAILED: 'Enrollment.SaveFailed',
 };
 
 export const setCommonEnrollmentSiteData = (enrollment: ApiEnrollment, attributeValues: ApiAttributeValues) =>
@@ -27,3 +31,21 @@ export const commitEnrollmentEvent = (eventId: string) =>
     actionCreator(enrollmentSiteActionTypes.COMMIT_ENROLLMENT_EVENT)({
         eventId,
     });
+
+export const updateEnrollmentEventsWithoutId = (uid: string, eventData: Object) =>
+    actionCreator(enrollmentSiteActionTypes.UPDATE_ENROLLMENT_EVENTS_WITHOUT_ID)({
+        eventData,
+        uid,
+    });
+
+export const rollbackEnrollmentEventWithoutId = (uid: string) =>
+    actionCreator(enrollmentSiteActionTypes.ROLLBACK_ENROLLMENT_EVENT_WITHOUT_ID)({
+        uid,
+    });
+
+export const commitEnrollmentEventWithoutId = (uid: string, eventId: string) =>
+    actionCreator(enrollmentSiteActionTypes.COMMIT_ENROLLMENT_EVENT_WITHOUT_ID)({
+        eventId,
+        uid,
+    });
+export const saveFailed = () => actionCreator(enrollmentSiteActionTypes.SAVE_FAILED)();