+++ src/core_modules/capture-core/components/WidgetEnrollmentEventNew/AccessVerification/accessVerification.types.js
@@ -13,6 +13,7 @@ export type ContainerProps = {|
     onSaveExternal?: ExternalSaveHandler,
     onSaveSuccessActionType?: string,
     onSaveErrorActionType?: string,
+    widgetReducerName: string,
     onCancel?: () => void,
 |};