+++ src/core_modules/capture-core/components/WidgetEnrollmentEventNew/Validated/validated.actions.js
@@ -51,7 +51,7 @@ export const requestSaveEvent = ({
         onSaveErrorActionType,
     }, { skipLogging: ['formFoundation'] });
 
-export const saveEvent = (serverData: Object, onSaveSuccessActionType?: string, onSaveErrorActionType?: string) =>
+export const saveEvent = (serverData: Object, onSaveSuccessActionType?: string, onSaveErrorActionType?: string, uid: string) =>
     actionCreator(newEventWidgetActionTypes.EVENT_SAVE)({}, {
         offline: {
             effect: {
@@ -59,7 +59,7 @@ export const saveEvent = (serverData: Object, onSaveSuccessActionType?: string, 
                 method: effectMethods.POST,
                 data: serverData,
             },
-            commit: onSaveSuccessActionType && { type: onSaveSuccessActionType, meta: { serverData } },
-            rollback: onSaveErrorActionType && { type: onSaveErrorActionType, meta: { serverData } },
+            commit: onSaveSuccessActionType && { type: onSaveSuccessActionType, meta: { serverData, uid } },
+            rollback: onSaveErrorActionType && { type: onSaveErrorActionType, meta: { serverData, uid } },
         },
     });