+++ src/core_modules/capture-core/components/WidgetEnrollmentEventNew/WidgetEnrollmentEventNew.types.js
@@ -12,4 +12,5 @@ export type WidgetProps = {|
     onSaveSuccessActionType?: string,
     onSaveErrorActionType?: string,
     onCancel?: () => void,
+    widgetReducerName: string
 |};