+++ src/core_modules/capture-core/components/WidgetStagesAndEvents/Stages/Stage/StageDetail/hooks/helpers.js
@@ -1,7 +1,7 @@
 // @flow
 import React from 'react';
 import moment from 'moment';
-import { statusTypes, translatedStatusTypes } from '../../../../../../metaData';
+import { statusTypes, translatedStatusTypes } from 'capture-core/events/statusTypes';
 import { convertMomentToDateFormatString } from '../../../../../../utils/converters/date';
 import { getSubValues } from '../../getEventDataWithSubValue';
 import type { StageDataElement } from '../../../../types/common.types';
@@ -18,6 +18,10 @@ const getEventStatus = (event: ApiEnrollmentEvent) => {
 
     if (isEventOverdue(event)) {
         return { status: statusTypes.OVERDUE, options: daysUntilDueDate ? dueDateFromNow : undefined };
+    }
+    // DHIS2-11576: VISITED status is treated as ACTIVE
+    if (event.status === 'VISITED') {
+        return { status: statusTypes.ACTIVE, options: undefined };
     }
 
     if (event.status === statusTypes.SCHEDULE) {