+++ src/core_modules/capture-core/reducers/descriptions/enrollmentDomain.reducerDescription.js
@@ -1,13 +1,18 @@
 // @flow
 import { createReducerDescription } from '../../trackerRedux/trackerReducer';
 import { enrollmentSiteActionTypes } from '../../components/Pages/common/EnrollmentOverviewDomain';
+import { actionTypes as enrollmentNoteActionTypes }
+    from '../../components/WidgetEnrollmentComment/WidgetEnrollmentComment.actions';
 
 const initialReducerValue = {};
 const {
     COMMON_ENROLLMENT_SITE_DATA_SET,
     UPDATE_ENROLLMENT_EVENTS,
+    UPDATE_ENROLLMENT_EVENTS_WITHOUT_ID,
     ROLLBACK_ENROLLMENT_EVENT,
+    ROLLBACK_ENROLLMENT_EVENT_WITHOUT_ID,
     COMMIT_ENROLLMENT_EVENT,
+    COMMIT_ENROLLMENT_EVENT_WITHOUT_ID,
 } = enrollmentSiteActionTypes;
 
 export const enrollmentDomainDesc = createReducerDescription(
@@ -56,6 +61,29 @@ export const enrollmentDomainDesc = createReducerDescription(
 
             return { ...state, events };
         },
+        [UPDATE_ENROLLMENT_EVENTS_WITHOUT_ID]: (
+            state,
+            { payload: { eventData, uid } },
+        ) => {
+            const events = [...state.events, { ...eventData, uid, pendingApiResponse: true }];
+            return { ...state, events };
+        },
+        [ROLLBACK_ENROLLMENT_EVENT_WITHOUT_ID]: (state, { payload: { uid } }) => {
+            const events = state.events?.filter(event => (event.uid !== uid));
+            return { ...state, events };
+        },
+        [COMMIT_ENROLLMENT_EVENT_WITHOUT_ID]: (state, { payload: { eventId, uid } }) => {
+            const events = state.events?.map((event) => {
+                if (event.uid === uid) {
+                    const { pendingApiResponse, uid: uidToRemove, ...dataToCommit } = event;
+                    return { ...dataToCommit, event: eventId };
+                }
+                return event;
+            });
+            return { ...state, events };
+        },
+        [enrollmentNoteActionTypes.ADD_ENROLLMENT_NOTE]:
+        (state, { payload: { note } }) => ({ ...state, notes: [...state.notes, note] }),
     },
     'enrollmentDomain',
     initialReducerValue,